Split NPC class into multiple sub classes. Added new Raccon NPC. Added food item class.
Player drops food item when hit by npc.
Player has a cooldown priod when hit by an npc.
Added Npc States.
Npcs prioritise food items instead of player.

Fixed Collision detection on food items.
Fixed multiple issuses with rendering sprites and animated sprites.

+++ TheRedPicnic/Framework/menu.h
@@ -30,8 +30,8 @@ public:
 	int GetSelected() const;
 	Button& GetSelectedButton() const;
 
-	int GetSelectionCooldown();
-	void SetSelectionCooldown(int cooldown);
+	float GetSelectionCooldown();
+	void SetSelectionCooldown(float cooldown);
 
 	void PositionElements(Vector2f containerDimensions);
 
@@ -41,7 +41,7 @@ private:
 
 	//Member Data:
 public:
-	int m_SelectionCooldown;
+	float m_SelectionCooldown;
 
 protected:
